requirements.txt
Flask==2.1.2
google-cloud-storage==2.5.0
Werkzeug==2.1.2
psycopg2-binary==2.9.7
python-dotenv==1.0.1
flask-cors==5.0.0
google-cloud-pubsub
